Midway through redoing ship/weapon menu/prefabs
Set up a new test prefab ship and weapon. Still need to instantiate weapons when levels start based on the WeaponInfo class. Also need to set up buy menu for weapons. And of course need to make a few more prefabs of each. Going for 6 ships and 10 weapons tops. Still might be some big logical issues with this system as is, with storing currentShipPrefab in an SO but having a changable WeaponInfo class? Should change SO to hold instance instead? Or something else entirely?